fix: long truncate width in flink 1.19 and 1.20

+++ flink/v1.20/flink/src/main/java/org/apache/iceberg/flink/sink/BucketPartitionerUtil.java
@@ -87,7 +87,7 @@ final class BucketPartitionerUtil {
 
     @Override
     public Tuple2<Integer, Integer> truncate(
-        int fieldId, String sourceName, int sourceId, int width) {
+        int fieldId, String sourceName, int sourceId, long width) {
       return null;
     }
